input: edt-ft5x06: Only read data for number of points reported
[Commit 462da817e6dac9e1fda65f7258ddce325ab9d4f3 rpi-linux]

Rather than always reading the maximum number of points supported
by the chip (which may be as high as 10), read the number of
active points first, and read data for just those.
In most cases this will result in less data on the I2C bus,
with only the maximum touch points taking more due to a second
read that has to configure the start address.

Patch Link:
https://github.com/raspberrypi/linux/commit/462da817e6dac9e1fda65f7258ddce325ab9d4f3

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>