drm/panel: Add prepare_upstream_first flag to drm_panel
[Commit f0f8bab8cb7c50cd1fbcb36854fe138c384f0495 rpi-linux]
[Commit da3e99addd7f2a59d9a7c94bd473818bb6ae959a rpi-linux]

Mapping to the drm_bridge flag pre_enable_upstream_first,
add a new flag prepare_upstream_first to drm_panel to allow
the panel driver to request that the upstream bridge should
be pre_enabled before the panel prepare.

This commit is a squash of the following 2 commits.

Patch Links:
https://github.com/raspberrypi/linux/commit/f0f8bab8cb7c50cd1fbcb36854fe138c384f0495
https://github.com/raspberrypi/linux/commit/da3e99addd7f2a59d9a7c94bd473818bb6ae959a

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>