arm64: dts: ti: k3-am62p-main: Add DSS1, DSI and DPHY-TX
TI's AM62P SoC supports a DPI to DSI video signal conversion bridge on
it's platform bus. The IP is from Cadence, and it has a custom TI
wrapper around it to facillitate integration.

This IP takes the DPI video signals from DSS1, and alongwith the DPHY IP,
it transmits DSI video signals out of the SoC.

Add support for DSS1, DSI bridge and the DPHY-TX.

Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>