firmware: ti_sci: Do not expect a response for entering partial-io
The current implementation uses the same code as
ti_sci_cmd_prepare_sleep() which expects a response to the prepare sleep
request. However when entering Partial-IO, there will not be any
response.

This patch rewrites tisci_enter_partial_io() to use a pure send function
without waiting for a response. After sending the prepare_sleep command,
we are expecting the system to enter Partial-IO immediately. If there is
an error before the send was successful, we report it with an error
message.

Signed-off-by: Markus Schneider-Pargmann <msp@baylibre.com>