HACK: configs: ti_rt: remove more kernel features with latency effects
Unless you have a looming deadline and a very aggressive latency goal
for the real-time Linux configuration you will most likely want to
keep these options enabled.

These options have the potential to do a lot to protect the kernel if a
vulnerability is found. Things like randomizing the base address of the
kernel to deter exploits requiring knowledge of the location of the
kernel, or guard pages and canaries to protect against stack smashing.

This also removes the boosting of the RCU thread to prevent it from
competing with the actual real-time task we care about.

Though these changes do have an impact on latency on low powered
devices. Testing has shown these changes remove around 15us.

All of this to say, I'm labeling this a particularly abominable hack
that will only motivate me to explore better options later. :)

Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Bryan Brattlof <bb@ti.com>