firmware: ti_sci: Use while loop after entering partial-io
machine_halt() is an internal function and can't be used from a module.

The driver is listed as tristate in Kconfig but selected as built-in
from ARCH_K3. However it is possible to build it as a module and
Android seems to do so as well.

machine_halt() can safely be replaced with a while (1) loop as there is
only one CPU active at this point. In case of an error we just need to
wait for the watchdog to reset the system.

Signed-off-by: Markus Schneider-Pargmann <msp@baylibre.com>
Reviewed-by: Dhruva Gole <d-gole@ti.com>