HACK: arm64: dts: ti: Configure am62p pinctrl-single for wake-up interrupts
Configure am62p pinctrl-single for wake-up interrupts.

Based on the patch for am625. No changes to the length of the register
set and the interrupt ID were needed for am62p as they match those on
am625.

Marked as HACK because physically none of the wakeup source events are
connected to GIC. Eventually we may want to set up a chained interrupt
handler in the ti-msgmgr for pinctrl-single to use.

Signed-off-by: Vibhore Vardhan <vibhore@ti.com>