arm64: dts: ti: k3-j784s4-main: Fix MAIN CPSW9G MAC Ports 5 to 8
The "ti,mac-only" property is used to indicate whether a port is
configured as a MAC Only port or not. Adding the property in the
device-tree node corresponding to a particular MAC Port specifies that
the port will operate in MAC mode of operation. The absence of the
property specifies that the group of ports which do not have the
property set will form a Switch thereby forwarding all broadcast traffic
amongst them in Hardware. Since the intended mode of operation on
device boot up is MAC mode of operation for all CPSW9G ports, add the
missing "ti,mac-only" property for MAC ports 5 to 8.

Fixes: 8bb677628e85 ("arm64: dts: ti: k3-j784s4-main: Add CPSW9G nodes")
Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>