config: Add config to enable faster boot for display
This disables configs that add to the kernel boot time and are
unnecessary for early boot. It also converts many features into modules
so that functionality is not restricted and can be loaded at run-time.
All configs that cannot be made loadable have to be removed.

DRM and TIDSS drivers are to be built as part of the Image instead of
loading them as modules.

The user is expected to understand the functionality that comes along
with removing these configs and enable them at the cost of boot time.

Major subsystems that will be removed in this config: SPI, USB_SUPPORT,
MMC, PWM, MTD_RAW_NAND

Usage: This config is run on top of defconfig and ti_arm64_prune.config

Signed-off-by: Aashvij Shenai <a-shenai@ti.com>
Reviewed-by: Jai Luthra <j-luthra@ti.com>
Reviewed-by: Aradhya Bhatia <a-bhatia1@ti.com>
Reviewed-by: Dhruva Gole <d-gole@ti.com>