mmc: sdhci_am654: Fix comments in set_clock functions
The sdhci_am654_set_clock function is also used to enable
delay chain, therefore fix comment to be more generic in
case we are not enabling the DLL.

Fix comment in sdhci_j721e_4bit_set_clock function, since
DLL is not enabled in this function. Add comment for input
tap delay while we are at it.

Fixes: fe52e2fbc6ef ("mmc: sdhci_am654: Fix conditions for enabling dll")

Signed-off-by: Judith Mendez <jm@ti.com>