net: ethernet: ti: cpsw-proxy-client: Fix channel validity check in detach
In the "cpsw_proxy_client_detach()" function, the validity of the TX and RX
DMA Channels is supposed to be checked in order to notify EthFw that they
are being released and no longer used. However, the existing implementation
incorrectly validates the address of the "is_valid" member of the TX and RX
channel structures rather than its value. Fix this.

Fixes: 5f1b32f6bd1b ("net: ethernet: ti: cpsw-proxy-client: Switch to ATTACH Request")
Reported-by: kernel test robot <lkp@intel.com>
Reported-by: Francesco Dolcini <francesco@dolcini.it>
Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>