net: ethernet: ti: cpsw-proxy-client: Drop unused "response_status"
The variable "response_status" in "cpsw_proxy_client_cb()" stores the
status of the response message received from EthFw. However the status
of the response is not checked in the "cpsw_proxy_client_cb()" function.
Rather, it is checked in the "cpsw_proxy_client_send_request()" function
which sends a request to EthFw and processes the response passed to it via
the "cpsw_proxy_client_cb()" callback function. Thus drop this unused
"response_status" variable in the callback function.

Fixes: 9343810f28a1 ("net: ethernet: ti: Add CPSW Proxy Client driver")
Reported-by: kernel test robot <lkp@intel.com>
Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>