net: ethernet: ti: icss_iep: Move the declarations to .h file
With the declarations of struct and enum related to IEP
present in .c file, it is not possible to access the structures
in other source code files. So move the declarations from
icss_iep.c to icss_iep.h.

IEP's def_inc is derived from the IEP reference clock.
Instead of using hard coded value, use the def_inc as
this is safer when changing the reference clock value.

Signed-off-by: Ravi Gunasekaran <r-gunasekaran@ti.com>