dmaengine: ti: k3-udma: Fix teardown for cyclic RX with PDMA
When receiving data in cyclic mode from PDMA peripherals, where reload
count is set to infinite, any TR in the set can potentially be the last
one of the overall transfer. In such cases, the EOP flag needs to be set
in each TR and PDMA's Static TR "Z" parameter should be set, matching
the size of the TR.

This is required for the teardown to function properly and cleanup the
internal state memory. This only affects platforms using BCDMA and not
those using UDMA-P, which could set EOP flag in the teardown TR
automatically.

Signed-off-by: Jai Luthra <j-luthra@ti.com>
Reviewed-by: Kamlesh Gurudasani <kamlesh@ti.com>