dt-bindings: gpu: ti,rogue: add clocks and power domains
Add the clocks and add extra power domains for the other BXS devices.
This allows users to set the clock speed for all BXS devices in
device-tree as opposed to userspace or driver specific overrides.

This also adds the dust power domain to these devices, bringing the
control of these domains back into the driver. Firmware controlled APM
would only add extra overhead to the current OS APM as we'd have to
structure domain interactions such that dust is pulled down before the
OS pulls down the firmware domain.

Signed-off-by: Randolph Sapp <rs@ti.com>