drm/bridge: cdns-dsi: Fix the link and phy init order
The order of init of DSI link and DSI phy is wrong. The DSI link needs
to be configured before the DSI phy is getting configured. Otherwise,
the DPHY is unable to lock in on the incoming PLL Reference clock.

Refer J721E TRM[0] section 12.6.5.7.3 "Start-up Procedure" for more
details.

Fix the order of inits.

[0]: J721E TRM
https://www.ti.com/lit/zip/spruil1

Fixes: e19233955d9e ("drm/bridge: Add Cadence DSI driver")
Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>