HACK: drm/panel: panel-ilitek9881: Add media bus format for the panel
Add RGB888_1X24 as the bus format for the pane, for tidss to absorb it.

This isn't the right way to add. Ideally, the fixes go in cdns-dsi driver.
And to adhere to those fixes, tidss should be supporting
DRM_BRIDGE_ATTACH_NO_CONNECTOR, which it doesn't in ti-6.1 kernel (but
does so in upstream/ti-6.6 kernels).

Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>