dt-bindings: gpu: Add PowerVR Series5 SGX GPUs
[ Upstream commit 796da8ca7e058ecf7ed2021cb8ce32c2318f83c4 ]

This version is slightly different from upstream as it removes
references to files not backported for this branch.

The Imagination PowerVR Series5 "SGX" GPU is part of several SoCs from
multiple vendors. Describe how the SGX GPU is integrated in these SoC,
including register space and interrupts. Clocks, reset, and power domain
information is SoC specific.

Signed-off-by: Andrew Davis <afd@ti.com>
Reviewed-by: Javier Martinez Canillas <javierm@redhat.com>
Reviewed-by: Krzysztof Kozlowski <krzysztof.kozlowski@linaro.org>
Message-ID: <20240109171950.31010-3-afd@ti.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>
Signed-off-by: Randolph Sapp <rs@ti.com>