drm: bridge: cdns-mhdp8546-core: Move mode_valid hook to bridge_funcs
With the support for the 'DRM_BRIDGE_ATTACH_NO_CONNECTOR' case,
the connector_helper funcs are not initialized if the encoder has this
flag in its bridge_attach call. Till now we had mode_valid hook only in
the drm_connector_helper_funcs. Move this hook in drm_bridge_funcs to
validate the modes in both case based on the flag.

Signed-off-by: Jayesh Choudhary <j-choudhary@ti.com>