HACK: drm/bridge: ti-sn65dsi86: Disable ASSR
Disable ASSR Display authentication, since ASSR is only supported
in eDP and not supported in DP. So disabling ASSR to support DP.

Marking it as HACK as we are disabling ASSR unconditionally.
Proper fix will include binding change for sn65dsi86 to
conditionally enabling/disabling ASSR based on underlying
protocol supporting it or not.

Signed-off-by: Rahul T R <r-ravikumar@ti.com>
Signed-off-by: Jayesh Choudhary <j-choudhary@ti.com>