arm64: defconfig: Enable DP display bridge for AM654 GP EVM
Enable the Toshiba TC358767 DPI to DP bridge and the Cadence D-PHY that
is required to transmit the DP signals.

A DP-display add-on board sits on top of TI's AM654 GP EVM board that
converts the DPI signals from the video port 2 of the DSS to the DP
signals. It uses the Cadence D-PHY to transmit the converted signals.

Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>