dt-bindings: display: ti,am65x-dss: Add support for AM62P DSS
TI's AM62P SoC supports 2 Display SubSystem (DSS) controllers, DSS0 and
DSS1.

The DSS0 HW supports one each of video pipeline (vid) and video-lite
pipeline (vidl1). It outputs OLDI signals on one video port (vp1) and
DPI signals on another (vp2). The video ports are connected to the
pipelines via 2 identical overlay managers (ovr1 and ovr2).

The DSS1 HW is similar but the VP directs to slightly different outputs.
The vp1 of DSS1 can either output a single-link OLDI output or a DPI
output. The vp2 of DSS1 can either output the DPI output or can
connector to a DSI Tx internally, that can give out DSI video output
from the SoC.

This SoC has separate LPSCs for OLDI0 and OLDI1, so the maxinum number
of power-domains required to switch on is now 3.

Add the new compatible for DSS0 and DSS1, and increase the number of possible
power domains to 3.

Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>