firmware: ti_sci: re-assign suspend-resume functions as _noirq callbacks
According to the [0] pm driver-api documentation, when the system goes
into the freeze, standby or memory sleep state, the phases are: prepare,
suspend, suspend_late, suspend_noirq. Since the TI_SCI driver is
responsible for issuing a prepare_system_suspend where we map and
validate the target Linux suspend state to TISCI LPM, it's suspend
resume calls should be among the last one's to be made.

This callback method will also ensure that all necessary wake-up
enabling
is done beforehand such that the SOC is now ready to finally enter the
appropriate low-power state.

[0] https://www.kernel.org/doc/html/latest/driver-api/pm/devices.html

Signed-off-by: Dhruva Gole <d-gole@ti.com>