arm64: dts: ti: Add overlay to force max OPPs on AM62A7-SK Rev E3
This overlay forces maximum clock rates for A53 and C7x on AM62A7-SK
Rev E3 board. It is meant only to be used for boards that have VDD_CORE
at 0.85V. Please note that AM62A7-SK Board revisions E1 and E2 do not
meet this requirement.

The overlay adds the 1.4 GHz OPP for A53 cores and disables all the
lower OPPs. It also overrides the clock rate for C7x to 1 GHz.

Any other boards based on this design should verify that they have the
right silicon variant and the right power tree before booting with this
overlay.

Signed-off-by: Vibhore Vardhan <vibhore@ti.com>