arm64: dts: ti: k3-j721s2-main: SRAM Allocation for VPU
Wave5 has option to connect device to on-chip SRAM through secondary AXI.
This allows for storing temporal buffers on a per stream basis leading to
the reduction of memory traffic.

Add in SRAM node for main_navss as no other component is utilizing it. Most
of MSMC SRAM is being used for TI deeplearning cases.

Signed-off-by: Brandon Brnich <b-brnich@ti.com>