arm64: dts: ti: k3-am642-evm/sk: Add CPSW3G CPTS PPS support
The CPTS driver is capable of configuring GENFy (Periodic Signal Generator
Function) present in the CPTS module, to generate periodic output signals
with a custom time period. In order to generate a PPS signal on the GENFy
output, the device-tree property "ti,pps" has to be used. The "ti,pps"
property is used to declare the mapping between the CPTS HWx_TS_PUSH
(Hardware Timestamp trigger) input and the GENFy output that is configured
to generate a PPS signal. The mapping is of the form:
<x-1 y>
where the value x corresponds to HWx_TS_PUSH input (1-based indexing) and
the value y corresponds to GENFy (0-based indexing).

To verify that the signal is a PPS signal, the GENFy output signal is fed
into the CPTS HWx_TS_PUSH input, which generates a timestamp event on the
rising edge of the GENFy signal. The GENFy output signal can be routed to
the HWx_TS_PUSH input by using the Time Sync Router. This is done by
mentioning the mapping between the GENFy output and the HWx_TS_PUSH input
within the "timesync_router" device-tree node.

The Input Sources to the Time Sync Router are documented at: [0]
The Output Destinations of the Time Sync Router are documented at: [1]

The ptp clock corresponding to the am65-cpsw-nuss driver for the MCU
CPSW instance has to be used to test PPS signal generation.

Use the command:
 # readlink -f /sys/class/ptp/ptp*/device/driver
in order to identify the ptp clock instance corresponding to the
am65-cpsw-nuss driver.

Enable PPS signal generation on GENF1. Use HW8_TS_PUSH Timestamp trigger
input to verify the PPS signal.

The PPS signal can be verified using testptp and ppstest tools as follows:
 # ./testptp -d /dev/ptp0 -P 1
 pps for system time request okay
 # ./ppstest /dev/pps0
 trying PPS source "/dev/pps0"
 found PPS source "/dev/pps0"
 ok, found 1 source(s), now start fetching data...
source 0 - assert 48.000000011, sequence: 4 - clear  0.000000000, sequence: 0
source 0 - assert 49.000000011, sequence: 5 - clear  0.000000000, sequence: 0
source 0 - assert 50.000000011, sequence: 6 - clear  0.000000000, sequence: 0
source 0 - assert 51.000000011, sequence: 7 - clear  0.000000000, sequence: 0
source 0 - assert 52.000000011, sequence: 8 - clear  0.000000000, sequence: 0

[0]
Link: https://software-dl.ti.com/tisci/esd/latest/5_soc_doc/am64x/interrupt_cfg.html#timesync-event-introuter0-interrupt-router-input-sources
[1]
Link: https://software-dl.ti.com/tisci/esd/latest/5_soc_doc/am64x/interrupt_cfg.html#timesync-event-introuter0-interrupt-router-output-destinations

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Signed-off-by: Chintan Vankar <c-vankar@ti.com>