arm64: dts: ti: k3-j721e: Add overlay to use INA2xx power monitors
Add an overlay to support using the INA2xx power monitors to measure
power from Linux.

While it is possible to move this support to the Common Processor Board
DTB, keep it as an overlay so that users who want to access the power
monitors via external connections such as using an automation header;
can continue to do so without undoing Linux DT nodes.

Signed-off-by: Keerthy <j-keerthy@ti.com>
Signed-off-by: Neha Malcom Francis <n-francis@ti.com>