net: ethernet: ti: cpsw-proxy-client: enable reset-recovery support
In the event of Hardware Error, EthFw is capable of detecting the Error
and notifying all Clients of this error. Once the Clients handle the
Error notification from EthFw and notify the same to EthFw, EthFw begins
the recovery sequence for CPSW. Using the existing helpers for handling
the error and recovery completion notifications from EthFw, enable the
Client driver support for reset-recovery sequence.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>