media: vxe-vxd: encoder: Fix Buffer Alignment of Encoder buffer
The Non 16 multiple resolutions were rounded up to lower 16 multiple, and so the
buffers were also getting rounded up to lower 16 multiple.
This make the output resolution also getting lower than expected resolution.
This patch fixes the buffer alignment for resolution which are non-multiple of 16,
ceil to multiple of 16 and set crop parameters used to parsei correct resolution
by decoder for Non 16 multiple resolutions. Alignment is done with respect to the
macro MB_SIZE defined to 16.

Signed-off-by: Prasanth Babu Mantena <p-mantena@ti.com>
Signed-off-by: Darren Etheridge <detheridge@ti.com>
Signed-off-by: Brandon Brnich <b-brnich@ti.com>