arm64: defconfig: Enable MTD support for SPI NAND flash
SPI NAND flashes are common in ARM64 SoCs. MTD SPI NAND
is the subsystem that drives these flashes. K3 TI SoCs
also use MTD SPI NAND support, enable it in defconfig as
built-in because it enables the NAND CORE and SPI MEM which
probes the SPINAND driver and enables the flash ops.

Signed-off-by: Apurva Nandan <a-nandan@ti.com>
Signed-off-by: Santhosh Kumar K <s-k6@ti.com>