dt-bindings: mfd: syscon: add TI's opp table compatible
The JTAG_USER_ID_USERCODE efuse address, which is located inside the
WKUP_CTRL_MMR0 range holds information to identify the speed grades of
various components on TI's K3 SoCs. Add a compatible to allow the
cpufreq driver to obtain the data to limit the maximum frequency for the
CPUs under Linux control.

[bb@ti.com: dropped compatibles when backporting to ti-linux-6.6.y]
Signed-off-by: Bryan Brattlof <bb@ti.com>