arm64: dts: ti: k3-am625-beagleplay: Reserve 128MiB of global CMA
Reserve 128MiB of global CMA which is also marked as re-usable so that OS
can also use the same if peripheral drivers are not using same.

AM62x supports multimedia components such as GPU, dual Display and ra.
Assuming the worst-case scenario where all 3 are run in parallel w is the
calculation :

1) OV5640 camera sensor supports 1920x1080 resolution -> 1920 width x 1080
   height x 2 bytesperpixel x 8 buffers (default in yavta) : 32MiB

2) 1920x1200 Microtips LVDS panel supported -> 1920 width x 1080 height x 4
   bytesperpixel x 2 buffers :
   16 MiB

3) 1920x1080 HDMI display supported -> 1920 width x 1080 height x 4
   bytesperpixel x 2 buffers :
   15.82 MiB which is ~16 MiB

4) IMG GPU shares with display allocated buffers while rendering but in
   case some dedicated operation viz color conversion, keeping same window
of ~16 MiB for GPU too.

Total is 80 MiB and adding 32 MiB for other peripherals and extra
16 MiB to keep as buffer for fragmentation thus rounding total to 128 MiB.

Signed-off-by: Devarsh Thakkar <devarsht@ti.com>
Tested-by: Chirag Shilwant <c-shilwant@ti.com>