arm64: dts: ti: k3-am625-beagleplay: Add Grove connector pinmux options
The Grove connector pins on BeaglePlay supports 3 different functions.
Describe these here.

Also move the default to be selected by a new Grove connector node.
As I2C is not the only possible function, do not enable main_i2c1.
It should only be enabled when a add-on board is connected that
makes use of it. That enabling should be done by the add-on board's
overlay.

Signed-off-by: Andrew Davis <afd@ti.com>