media: img: vxe-vxd: fix incomplete decoder initialization
With Scarthgap Gstreamer, the s_fmt is no longer called on the CAPTURE
port of the v4l2 device.  The s_fmt needs to be called in this driver to
set a whole bunch of decoder state information.  If it is not called the
state information is not configured and the decoder very quickly errors
out when trying to decode frmaes as it has no idea the pixel format, or
the dimensions of the CAPTURE buffer.

This fix looks to see if the decoder state matches what we are going to
report back to the v4l2 framework, and if it doesn't it calls an
explicit s_fmt to set the correct context information for the decoder.

Signed-off-by: Darren Etheridge <detheridge@ti.com>
Reviewed-by: Brandon Brnich <b-brnich@ti.com>