remoteproc: k3-r5: keep the device on during stop
Keep the core on when stopping a R5F core. On a stop, the core is
halted and held in reset. On a start, the reset is released after a
firmware loaded and this boots the core.
This mechanism allows each R5F core in a cluster to be
independently stopped and loaded with a new firmware.

Signed-off-by: Hari Nagalla <hnagalla@ti.com>
Reviewed-by: Beleswar Padhi <b-padhi@ti.com>
Tested-by: Beleswar Padhi <b-padhi@ti.com>