v4l2-controls: Add Control for Background Detection
Background detection is a useful technique in video encoding to preserve
bits in cases when the video being encoded has a background that doesn't
change often. A predominant use case for this is security scenarios where
the background hardly changes. Some video IPs offer this feature as part of
their hardware. Expose this control so end-users can select this feature if
offered by the IP.

Signed-off-by: Nas Chung <nas.chung@chipsnmedia.com>
Signed-off-by: Brandon Brnich <b-brnich@ti.com>