TI: HACK: media: i2c: imx390: Remove V4L2_SUBDEV_FL_STREAMS flag to fix probe failure
After kernel v6.6, a check was introduced that enforces implementation
of the .enable_streams() and .disable_streams() callbacks
when the V4L2_SUBDEV_FL_STREAMS flag is set.

The IMX390 driver currently sets this flag but does not implement
the required APIs, causing probe failure.

This patch removes the V4L2_SUBDEV_FL_STREAMS flag as a temporary
workaround to allow the driver to function on v6.12.

A clean solution will involve implementing the .enable_streams()
and .disable_streams() APIs in the future.

Signed-off-by: Yemike Abhilash Chandra <y-abhilashchandra@ti.com>