FROMLIST: remoteproc: k3-dsp: Drop check performed in k3_dsp_rproc_{mbox_callback/kick}
Commit 8bdac02c5bfe ("remoteproc: k3-dsp: Acquire mailbox handle during
probe routine") introduced a check in the
"k3_dsp_rproc_mbox_callback()" and "k3_dsp_rproc_kick()" callbacks,
causing them to exit if the remote core's state is "RPROC_DETACHED".
However, the "__rproc_attach()" function that is responsible for
attaching to a remote core, updates the state of the remote core to
"RPROC_ATTACHED" only after invoking "rproc_start_subdevices()".

The "rproc_start_subdevices()" function triggers the probe of the Virtio
RPMsg devices associated with the remote core, which require that the
"k3_dsp_rproc_kick()" and "k3_dsp_rproc_mbox_callback()" callbacks are
functional. Hence, drop the check in the callbacks.

Fixes: 8bdac02c5bfe ("remoteproc: k3-dsp: Acquire mailbox handle during probe routine")
Link: https://lore.kernel.org/all/20250108063727.1416324-3-b-padhi@ti.com/
Signed-off-by: Beleswar Padhi <b-padhi@ti.com>