FROMLIST: remoteproc: k3: Refactor .start rproc ops into common driver
The k3_dsp_rproc_start() function sets the boot address and releases the
reset on the remote processor. Whereas, the k3_m4_rproc_start() function
only needs to release the reset. Refactor the k3_m4_rproc_start() into
ti_k3_common.c as k3_rproc_start() and align the DSP and M4 drivers to
invoke this common function when releasing the reset on the remote
processor.

Further, do not override the .start ops with NULL when operating in
IPC-only mode in ti_k3_dsp_remoteproc.c, as .start is never invoked in
rproc_attach routine.

Link: https://lore.kernel.org/all/20250103101231.1508151-12-b-padhi@ti.com/
Signed-off-by: Beleswar Padhi <b-padhi@ti.com>