PENDING: media: chips-media: wave5: Move hrtimer_start/cancel
Introducing runtime pm into the Wave5 driver has caused issue with hrtimer
functionality. The worker thread is executing when device is off which is
leading to driver hang up. Moving start/stop to runtime pm hooks addresses
this issue while also making hrtimer stop when Wave5 is not in use.

Signed-off-by: Brandon Brnich <b-brnich@ti.com>