TI: HACK: arm64: dts: ti: Configure am62p pinctrl-single for wake-up interrupts
Configure am62p pinctrl-single for wake-up interrupts.

Based on the patch for am625. No changes to the length of the register
set and the interrupt ID were needed for am62p as they match those on
am625.

Marked as HACK because physically none of the wakeup source events are
connected to GIC. Eventually we may want to set up a chained interrupt
handler in the ti-msgmgr for pinctrl-single to use.

[k-willis@ti.com] appended compatible with single-pinctrl to fix DT
warning. Squashed relevant  part of commit eb3b9e22be86 ("arm64: dts: ti:
k3-am62*: Append pinctrl-single compatible").

Signed-off-by: Vibhore Vardhan <vibhore@ti.com>
Signed-off-by: Dhruva Gole <d-gole@ti.com>
Signed-off-by: Devarsh Thakkar <devarsht@ti.com>
Reported-by: kernel test robot <lkp@intel.com>
Reviewed-by: Dhruva Gole <d-gole@ti.com>
Signed-off-by: Kendall Willis <k-willis@ti.com>