PENDING: arm64: dts: ti: k3-am62p-j722s-common-wakeup: Configure ti-sysc for wkup_uart0
Similar to am62-wakeup, am62p-j722s-common-wakeup is capable of
wakeup from wkup_uart0. ti-sysc manages the sysconfig registers while
wkup_uart0 is only a portion of the register space.

See the link below for the upstream version of this patch for AM62.

Built and tested on AM62p5-sk board.

[k-willis@ti.com] left the wkup_conf as bus node instead of change to be
syscon node as in original patch.

Link: https://lore.kernel.org/r/20231219072503.12427-1-tony@atomide.com

Signed-off-by: Vibhore Vardhan <vibhore@ti.com>
Signed-off-by: Kendall Willis <k-willis@ti.com>