FROMLIST: drm: bridge: cdns-mhdp8546: Move mode_valid hook to drm_bridge_funcs
With the support for the 'DRM_BRIDGE_ATTACH_NO_CONNECTOR' case,
the connector_helper funcs are not initialized if the encoder has this
flag in its bridge_attach call. Till now we had mode_valid hook only in
the drm_connector_helper_funcs. Move this hook to drm_bridge_funcs to
validate the modes.

Link: https://lore.kernel.org/all/20240530091757.433106-1-j-choudhary@ti.com/
Signed-off-by: Jayesh Choudhary <j-choudhary@ti.com>
Acked-by: Maxime Ripard <mripard@kernel.org>