PENDING: input: edt-ft5x06: Include I2C details in names for the devices
[commit e12af83fea018fd672140a2b77b6d96902112097 rpi-linux]

libinput uses the input device name alone. If you have two
identical input devices, then there is no way to differentiate
between them, and in the case of touchscreens that means no
way to associate them with the appropriate display device.

Add the I2C bus and address to the start of the input device
name so that the name is always unique within the system.

Patch Link:
https://github.com/raspberrypi/linux/commit/e12af83fea018fd672140a2b77b6d96902112097

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>
Signed-off-by: Devarsh Thakkar <devarsht@ti.com>