PENDING: arm64: dts: ti: k3-j784s4-main: add gpu node
Add the IMG BXS-4-64 GPU device node. This cannot be upstreamed as this
device currently uses an out-of-tree kernel module.

Add clock and power domains to the GPU node for j784s4. This brings the
dust power domain control back under the driver and allows users to
override the default clock rate for the GPU's core clock.

Signed-off-by: Antonios Christidis <a-christidis@ti.com>