PENDING: arm64: dts: ti: k3-j721e-main: add gpu node
Add the IMG 8XE-GE8430 GPU device node. This cannot be upstreamed as
this device currently uses an out-of-tree kernel module.

Add clock and power domains to the GPU node for J721E. This brings the
dust power domain control back under the driver and allows users to
override the default clock rate for the GPU's core clock.

Signed-off-by: Antonios Christidis <a-christidis@ti.com>