PENDING: arm64: dts: ti: k3-am642-evm/sk: Drop routing GenF to SYNC1_OUT signal
GenF was routed to SYNC1_OUT with the intent of having a way to measure
the PPS signal, since the board schematics indicated that SYNC1_OUT_TP
was connected to the J12 header. However, the datasheet indicates that
it is actually SYNC0_OUT and not SYNC1_OUT that is connected to the D18
pin which is routed to the J12 header. Hence, don't route GenF to
SYNC1_OUT since it doesn't serve the original purpose of providing a way
to measure the PPS signal.

Signed-off-by: Meghana Malladi <m-malladi@ti.com>