PENDING: arm64: dts: ti: k3-am642-evm: Add support to test ICSSG PPS using ppstest tool
In order to use open source ppstest tool which prints 64-bit
timestamp every time a pps event gets updated in the kernel,
the iep driver is responsible to register the pps event for every
pulse to get per second timestamp.

But as the firmware for IEP module doesn't support HW interrupt
or callback support for PPS, route IEP PPS to NAVSS CPTS via
TimeSync Router and use CPTS HW push events to generate events
per pulse, which will get registered in the kernel and update
the tool to print the pps timestamps.

Signed-off-by: Meghana Malladi <m-malladi@ti.com>