PENDING: drivers: crypto: sa2ul: Increase authentication context max size
SA2UL can have the max authentication context length of 160 bytes and
max encryption context length of 128 bytes, while creating the context,
we were always allocating some bytes but they weren't consistent across
the codebase and even with the variable saying it's just sending the
encryption context, the authentication context was being sent as well
with the SCCTL values.

Always tell SA2UL to fetch the encryption as well as authentication
context ( as we are doing today as well ) and just hardcode the max
sizes of the contexts that SA2UL can handle to make the code simpler.

Signed-off-by: Manorit Chawdhry <m-chawdhry@ti.com>