PENDING: can: m_can: Enable pinctrl wakeup without WoL enabled
In some devices the pins of the m_can module can act as a wakeup source.
am62 requires a wakeup flag being set in pinctrl when mcan pins acts as
a wakeup source. Add support to select the wakeup state on suspend.

Previously, m_can could only be a wakeup source when WoL is enabled.
This patch allows m_can to be a wakeup source regardless of WoL being
enabled.

Signed-off-by: Kendall Willis <k-willis@ti.com>