TI: HACK: remoteproc: k3-r5: do not powerdown r5f core on AM64
On AM64 devices, power-down of core1 in r5f cluster even when core0
is on, randomly gets the lpsc stuck in transition state and makes the core
un-usable. So, as a work around keep the core halted and reset asserted
during a stop. Since the cores in the r5f cluster are kept powered on after
the initial boot from por, relax the conditional logic to sequentially
stop/start cores in split mode of AM64 r5f cluster at runtime with sysfs
remote proc stop/start commands.

Signed-off-by: Hari Nagalla <hnagalla@ti.com>
Reviewed-by: Beleswar Padhi <b-padhi@ti.com>