TI: arm64: dts: ti: k3-am62: Remove MCU UART node from LPM overlay
Enabling MCU UART from Linux is causing issues in suspend path. MCU
UART is controlled by MCU firmware and Linux should not be controlling
it at the same time. Enabling MCU UART from Linux does not enable an
additional wakeup source for any LPM so there is no reason for it to be
in this overlay. For all these reasons, remove the MCU UART nodes from
this overlay.

Signed-off-by: Kendall Willis <k-willis@ti.com>