ARM: bb.org_defconfig: don't compile mmc_block as a module
As specified in the help for the MMC_BLOCK option: "This provides a
block device driver, which you can use to mount the filesystem."
The system is able to boot from MMC only if it is compiled as built-in
(by default, it is set to 'y').
If, instead, it is selected as a module, the system fails to complete
the boot process. The idea is that this module is needed to mount the
filesystem, but the module itself is located within the filesystem.

Signed-off-by: Dario Binacchi <dario.binacchi@amarulasolutions.com>